drivers: clk: use int64_t for clock frequencies

+++ include/tart/drivers/clk/rp2-osc.hpp
@@ -6,19 +6,19 @@
 namespace tart {
 
 struct rp2_xosc final : clk {
-	rp2_xosc(uintptr_t base, uint32_t freq)
+	rp2_xosc(uintptr_t base, int64_t freq)
 	: clk{nullptr}, freq_{freq}, space_{base} { }
 
 	virtual ~rp2_xosc() = default;
 
-	uint32_t freq() const override { return freq_; }
+	int64_t freq() const override { return freq_; }
 	void start() override;
 	void stop() override;
 
 	void busy_wait(uint8_t cycles);
 
 private:
-	uint32_t freq_;
+	int64_t freq_;
 
 	// TODO(qookie): Remove the :: prefix once tart::arch is gone.
 	::arch::mem_space space_;
